+++ src/OneDOF/Prog1/Undamped/CFrameAnimation.java
@@ -225,7 +225,7 @@ class CFrameAnimation extends CFrame {
       this.mTraceGC.fillRect(0, 0, 3280, 120);
       this.mTraceGC.setColor(Color.black);
 
-      for (byte var1 = 0; var1 < 3280; var1 += 50) {
+      for (int var1 = 0; var1 < 3280; var1 += 50) {
          this.mTraceGC.drawLine(var1, 0, var1, 120);
       }
